backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
mysqlclient==2.2.0
python-decouple==3.8
gunicorn==21.2.0
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0
